+++ packages/interface-compliance-tests/CHANGELOG.md
@@ -75,6 +75,25 @@
   * dependencies
     * @libp2p/utils bumped from ^5.2.4 to ^5.2.5
 
+## [6.3.7](https://github.com/libp2p/js-libp2p/compare/interface-compliance-tests-v6.3.6...interface-compliance-tests-v6.3.7) (2025-03-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.0.13 to ^5.0.14
+    * @libp2p/echo bumped from ^2.1.14 to ^2.1.15
+    * @libp2p/interface-internal bumped from ^2.3.5 to ^2.3.6
+    * @libp2p/logger bumped from ^5.1.10 to ^5.1.11
+    * @libp2p/memory bumped from ^1.1.1 to ^1.1.2
+    * @libp2p/multistream-select bumped from ^6.0.17 to ^6.0.18
+    * @libp2p/peer-collections bumped from ^6.0.21 to ^6.0.22
+    * @libp2p/peer-id bumped from ^5.0.14 to ^5.0.15
+    * @libp2p/plaintext bumped from ^2.0.17 to ^2.0.18
+    * @libp2p/utils bumped from ^6.5.5 to ^6.5.6
+    * libp2p bumped from ^2.7.4 to ^2.7.5
+
 ## [6.3.6](https://github.com/libp2p/js-libp2p/compare/interface-compliance-tests-v6.3.5...interface-compliance-tests-v6.3.6) (2025-02-26)